Create README.md
This is a basic thing that everyone can do! If anyone wish to learn Programming, I surely suggest to learn Scripting Language first! Because it will be a warm up for programmers. 

And this is a basic script that can do some destruction that we never expect! And please be careful when you handle this!